создал класс Transport, переопределил метод toString для класса, перенёс гетеры и сеттеры, переопределил toString для Car над этим ещё поработать нужно
