+++ src/pro/sky/java/course2/Lesson3_inheritanceOOP/transport/Car.java
@@ -3,61 +3,25 @@ package pro.sky.java.course2.Lesson3_inheritanceOOP.transport;
 import java.time.LocalDate;
 import java.util.Objects;
 
-public class Car {
-
-    /*Создайте пакет transport и перенесите туда класс Car.
-     Добавьте следующие свойства класса:
-     «Коробка передач»,
-     «Тип кузова»,
-     «Регистрационный номер»,
-     «Количество мест»,
-     признак «Летняя» или «Зимняя резина».*/
+public class Car extends Transport{
     private String transmission;
     private int registrationNumber;
     private final int numberOfSeats;
     private final String bodyType;
-    private final String brand;
-    private final String model;
     private boolean isSummerTyres;
     private double engineVolume;
-    private final int year;
-    private String color;
-    private final String country;
     private Key key;
     private static int countOfCars;
-    private static final String DEFAULT_VALUE = "default";
     private static final Integer DEFAULT_VALUE_OF_SEATS = 4;
     private static final double DEFAULT_ENGINE_VALUE = 1.5;
-    private static final String DEFAULT_COLOUR = "white";
     private static final String DEFAULT_TRANSMISSION = "auto";
 
-    public Car(String brand, String model, int year, String country, String color, double engineVolume, String transmission, String bodyType, int registrationNumber, int numberOfSeats, boolean isSummerTyres, Key key) {
+    public Car(Transport transport, double engineVolume, String transmission, String bodyType, int registrationNumber, int numberOfSeats, boolean isSummerTyres, Key key) {
         setRegistrationNumber(registrationNumber); //для оптимизации, вместо условий поставил сеттеры
         setTransmission(transmission);
         setKey(key);
-        setColor(color);
         setEngineVolume(engineVolume);
-
-        if (brand.isBlank() || brand == null) {
-            this.brand = DEFAULT_VALUE;
-        } else {
-            this.brand = brand;
-        }
-        if (model.isBlank() || model == null) {
-            this.model = DEFAULT_VALUE;
-        } else {
-            this.model = model;
-        }
-        if (year >= 0) {
-            this.year = year;
-        } else {
-            this.year = 2000;
-        }
-        if (country.isBlank() || country == null) {
-            this.country = DEFAULT_VALUE;
-        } else {
-            this.country = country;
-        }
+        transport = new Transport(); // чекнуть это на работоспособность, скорее всего запись должна быть другая
         if (bodyType.equals("Jeep") ||
                 bodyType.equals("Hatchback") || bodyType.equals("Sedan") || bodyType.equals("Coupe") ||
                 bodyType.equals("Wagon") || bodyType.equals("Van") || bodyType.equals("MUV/SUV") ||
@@ -77,8 +41,8 @@ public class Car {
         countOfCars++;
     }
 
-    public Car(String brand, String model, double engineVolume, int year) {
-        this(brand, model, year,"","", engineVolume, "",  "", 0, 0, false, new Key());
+    public Car(Transport transport, Key key) {
+        this(new Transport(), new Key());
         countOfCars++;
     }
 
@@ -90,28 +54,8 @@ public class Car {
         this.key = Objects.requireNonNullElseGet(key, Key::new);
     }
 
-    public String getBrand() {
-        return brand;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
     public double getEngineVolume() {
         return engineVolume;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public int getYear() {
-        return year;
-    }
-
-    public String getCountry() {
-        return country;
     }
 
     public String getTransmission() {
@@ -162,14 +106,6 @@ public class Car {
         }
     }
 
-    public void setColor(String color) {
-        if (color.isBlank() || color == null) {
-            this.color = DEFAULT_COLOUR;
-        } else {
-            this.color = color;
-        }
-    }
-
     public int getMonth() {
         return LocalDate.now().getMonthValue();
     }
@@ -180,9 +116,9 @@ public class Car {
 
     @Override
     public String toString() {
-        return '\n' + brand + " " + model + ", production year " +year+", made in " + country + ", colour is " + color + ", engine volume " + engineVolume + " l."
+        return  "engine volume " + engineVolume + " l."
                 + "\nnumbers of seats is " + numberOfSeats + ", body type " + bodyType + ", registration number " + registrationNumber + ", " + "\ntransmission "
-                + transmission + ", with " + (isSummerTyres? "summer tyres": "winter tyres");
+                + transmission + ", with " + (isSummerTyres? "summer tyres": "winter tyres"); //добавить transport и key
     }
 
     public static class Key {
